### 2019-07-07 #1546 @whedon
[PRE REVIEW]: MiraPy: Python Package for Deep Learning in Astronomy
**Submitting author:** @swapsha96 (<a href="http://orcid.org/0000-0002-0329-9314">Swapnil Sharma</a>)
**Repository:** <a href="https://github.com/mirapy-org/mirapy" target ="_blank">https://github.com/mirapy-org/mirapy</a>
**Version:** v0.1.0
