### 2019-07-07 #1546 @labarba
👋 @swapsha96 — Your paper is failing to compile. I think you need quotes around the title, or the colon confuses the key-value pair.

https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography
