### 2019-07-11 #1557 @mbobra
@swapsha96 I have a couple comments as editor.
* Can you please clarify how this submission differs from [AstroML](http://www.astroml.org/)?
* Can you please [cite Keras](https://keras.io/getting-started/faq/#how-should-i-cite-keras) in your paper?
* Can you please document the Jupyter notebooks in your tutorial section (at the moment they contain mostly code and no Markdown cells)? And write a short summary of each notebook and put that information in the tutorial section of the MiraPy Read The Docs site?
* Can you please document the functions? E.g. I don't know how `AtlasVarStarClassifier` differs from the native Keras classifiers and I don't know which input parameters (and what form) to pass into this function.

Please let me know if you want more clarification or any help with this.
