+++ paper.md
@@ -26,7 +26,7 @@ bibliography: paper.bib
 # Summary
 
 MiraPy is a Python package for Deep Learning in Astronomy. It is built using
-Keras for developing ML models to run on CPU and GPU seamlessly. The
+Keras [@chollet2015keras] for developing ML models to run on CPU and GPU seamlessly. The
 aim is to make applying machine learning techniques on astronomical data easy
 for astronomers, researchers and students.
 
