### 2019-07-11 f4e780a
fixed class docs
