### 2019-07-11 7da8c56
fixed class docs
