### 2019-07-11 443a445
fixed class docs in atlas classifier
