### 2019-07-11 fd718bc
add description in tutorials
