+++ docs/source/TUTORIALS.rst
@@ -1,4 +1,28 @@
 Tutorials
 =========
 
-You can find the Jupyter notebooks on various applications of MiraPy for problem-solving in Astronomy in our `Github repository <https://github.com/mirapy-org/tutorials>`_.
+We have a set of MiraPy tutorials for problem-solving in Astronomy using Deep Learning. You can find the Jupyter notebooks in our `Github repository <https://github.com/mirapy-org/tutorials>`_. Following are the short description of MiraPy applications:
+
+- **Astronomical Image Reconstruction using Autoencoder**
+
+    Encoder-decoder networks can be trained for noise removal from blurry image. We can use MiraPy for astronomical image reconstruction by training a simple denoising autoencoder using some images of galaxies and nebulae in Missier catalog.
+
+- **ATLAS variable star Classification**
+
+    We demonstrate how to use MiraPy to classify variable stars using features extracted from light curves. These features are available in ATLAS catalog. We use deep neural network for the same.
+
+- **OGLE variable star Classification**
+
+    We demonstrate how to use MiraPy to classify variable stars using light-curves available in OGLE variable star catalogs. We use Recurrent Neural Network (RNN) in the classification model.
+
+- **HTRU1 Batched Dataset Classification**
+
+    MiraPy can be used for the classification of pulsars and non-pulsars in dataset released by HTRU1 survey. The dataset contains 60000 images which are classified using Convolutional Neural Network (CNN).
+
+- **X-Ray Binary Classification**
+
+    Tutorial demonstrates how to use Fully-Connected Neural (FCN) network to classify features of pulsar, non-pulsar and black hole systems.
+
+- **2D and 3D visualisation**
+
+    We demonstrate how to use MiraPy to visualize a feature dataset using 2D and 3D graphs. For this purpose, we use Pricipal Component Analysis (PCA) for feature reduction.
