### 2019-07-11 #1557 @swapsha96
@mbobra Thank you for your input! I have made the changes to the best of my understanding. Please let me know what other changes I need to do.

- AstroML is built on top of scikit-learn which has design constraints that make it impossible to extend to deep learning techniques such as large sequence data (eg lightcurve) classification, CNN, and encoder-decoder networks. Importantly it does **not** support GPU which is very important for large matrix multiplications, processing large astronomical data, etc. Since a large amount of data is collected every day, we decided to create a new package that takes cares of that. We are developing it in such a way that not only astronomers can use MiraPy's existing features but they can extend its applications to their field of interests.

- I cited Keras in the paper. Thanks, I missed that.

- I have added markdown cells in all tutorial notebooks. Also, I have written a summary of all notebooks in the docs as per your suggestion.

[Tutorial repository](https://github.com/mirapy-org/tutorials)
[Tutorial section in doc](https://mirapy.readthedocs.io/en/latest/TUTORIALS.html)

- We had written it in `__init__` function instead of putting it under the class name which is why the documentation didn't show that. Apologies for that. I have fixed the problem now.

`AtlasVarStarClassifier` [doc](https://mirapy.readthedocs.io/en/latest/autoapi/mirapy/classifiers/models/index.html#mirapy.classifiers.models.AtlasVarStarClassifier)
