### 2019-07-12 1496c00
Add keras installation guide
