### 2019-07-20 mirapy-org/MiraPy#39 @DanielLenz
Adding tests
Hi, I'm currently reviewing `mirapy` for publication in JOSS (https://github.com/openjournals/joss-reviews/issues/1557).

Could you add a suite of (unit)tests to the project that verify the software? It's one of the JOSS requirements.

Thank you!
