### 2019-07-20 #1557 @DanielLenz
I've added a few points that should be completed before we can proceed with the submission.

The project is in great shape and I'm sure the authors could address these points quickly. However, I'm not sure if the scope of the project is sufficient: It's mostly tailored towards a very specific set of tasks, and is essentially a shallow wrapper around keras. @mbobra , could you comment and provide JOSS's policy on that?
