### 2019-07-21 #1557 @swapsha96
This software is written using specific examples but can be used to apply Deep Learning techniques in various other problems depending upon the use cases. For example, we experimented astronomical image reconstruction on just 38 images of some Messier's objects but it can be extended to images of galaxies from NASA's SDSS catalog and images captured by future James Webb Telescope. In tutorials, we demonstrated classification of variable stars using Lightcurves on OGLE dataset but the model can be applied on ATLAS variable star dataset and datasets from future projects.

MiraPy is in its initial stage and requires huge improved improvements and feedback. I hope to make this package more generalized in future.
