### 2019-07-21 mirapy-org/MiraPy#39 @swapsha96
Currently, I have written [tests](https://github.com/mirapy-org/mirapy/tree/master/mirapy/fitting/tests) for `fitting` module. Since other modules require large datasets for testing, I have not written any specific tests for them right now. You can run the tests by running the following command in the cloned repository:
```
python setup.py test
```

Since the requirement says:

>  Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

I hope this requirement is satisfied.
