### 2019-07-21 mirapy-org/MiraPy#39 @DanielLenz
Thanks for pointing that out! I agree that it's difficult to test parts of the package that require large data, but part of the JOSS requirements is that most of the codebase is tested, not only a small part of it.

It would be awesome if you could add some very simple unit tests, and possibly provide the test coverage as well.
