### 2019-07-22 1ba191a
add test for utils
