### 2019-07-22 097c658
Update scipy version
