### 2019-07-22 2797c79
Add coverage badge
