### 2019-07-22 c57d3a0
Update .travis.yml
