+++ .travis.yml
@@ -151,3 +151,4 @@ after_success:
     # If coveralls.io is set up for this package, uncomment the line below.
     # The coveragerc file may be customized as needed for your package.
     # - if [[ $SETUP_CMD == *coverage* ]]; then coveralls --rcfile='mirapy/tests/coveragerc'; fi
+    - bash <(curl -s https://codecov.io/bash)
