+++ mirapy/__init__.py
@@ -12,5 +12,16 @@ import sys
 
 __minimum_python_version__ = "3.5"
 
+
+class UnsupportedPythonError(Exception):
+    pass
+
+
+if sys.version_info \
+        < tuple((int(val) for val in __minimum_python_version__.split('.'))):
+    raise UnsupportedPythonError("mirapy does not support Python < {}"
+                                 .format(__minimum_python_version__))
+
 if not _ASTROPY_SETUP_:
     # For egg_info test builds to pass, put package imports here.
+    from mirapy.fitting import *
