+++ codecov.yml
@@ -1,0 +1,3 @@
+ignore:
+  - "mirapy/data/load_dataset.py"
+  - "_astropy_init.py"
