### 2019-07-22 87a67a2
add more tests
