### 2019-07-22 9943968
Update .travis.yml
