+++ codecov.yml
@@ -1,5 +1,1 @@
-ignore:
-  - "mirapy/data/load_dataset.py"
-  - "_astropy_init.py"
-  - "ah_bootstrap.py"
-  - "setup.py"
+
