### 2019-07-22 mirapy-org/MiraPy#39 @swapsha96
Currently, the test coverage is [43%](https://coveralls.io/github/mirapy-org/mirapy).
