### 2019-07-23 mirapy-org/MiraPy#39 @DanielLenz
Thanks for adding all these tests!
