+++ .travis.yml
@@ -145,12 +145,11 @@ install:
     # other dependencies.
 
 script:
-   - $MAIN_CMD $SETUP_CMD
-   - coverage run setup.py test
+   - $MAIN_CMD $SETUP_CMD --coverage
 
 after_success:
     # If coveralls.io is set up for this package, uncomment the line below.
     # The coveragerc file may be customized as needed for your package.
     # - if [[ $SETUP_CMD == *coverage* ]]; then coveralls --rcfile='mirapy/tests/coveragerc'; fi
     - bash <(curl -s https://codecov.io/bash)
-    - codecov
+    - coveralls
