### 2019-07-23 db6cb62
add data test and remove dataset
