### 2019-07-23 48c31b2
add messier catalog tests
