+++ mirapy/data/tests/test_load_dataset.py
@@ -1,5 +1,8 @@
 import os
 import pytest
+import numpy as np
+import scipy.misc
+from mirapy.utils import get_psf_airy
 from mirapy.data import load_dataset
 
 
@@ -16,3 +19,23 @@ def test_load_xray_binary_data():
     os.rmdir(path)
 
     assert len(x) == 1 and len(x) == len(y)
+
+def test_messier_catalog_images():
+    path = 'test_messier_catalog_image/'
+    filename = 'test.png'
+    img = np.zeros([100,100,3],dtype=np.uint8)
+    
+    os.mkdir(path)
+    scipy.misc.imsave(path+filename, img)
+
+    imgs = load_dataset.load_messier_catalog_images(path)
+
+    os.remove(path+filename)
+    os.rmdir(path)
+
+    assert len(imgs) == 1
+
+    psf = get_psf_airy(100, 2)
+    imgs, imgs_noisy = load_dataset.prepare_messier_catalog_images(imgs, psf, psf)
+
+    assert len(imgs) == len(imgs_noisy)
