### 2019-07-23 6632837
add cv2 in travis
