### 2019-07-23 5f3a163
add test for x-ray binary classification
