+++ mirapy/data/tests/test_load_dataset.py
@@ -12,7 +12,7 @@ def test_load_xray_binary_data():
     
     os.mkdir(path)
     with open(path+filename, 'w') as f:
-        f.write('J0352+309 Pulsar      50139.6     0.363137     0.995448      1.54628\n') 
+        f.write('J0352+309 Pulsar\t50139.6\t0.363137\t0.995448\t1.54628\n') 
     x, y = load_dataset.load_xray_binary_data(path)
     
     os.remove(path+filename)
