### 2019-07-23 #1557 @swapsha96
@mbobra @wtgee @DanielLenz The issues are now closed. What else do I need to look into?
