### 2019-07-23 f0cdc2c
remove few unimportant lines from mirapy/__init__.py
