### 2019-07-25 7c5db6a
Update README.rst
