+++ README.rst
@@ -9,7 +9,7 @@ MiraPy: Python Package for Deep Learning in Astronomy
     :target: http://www.astropy.org
     :alt: Powered by Astropy Badge
 
-.. image:: https://img.shields.io/travis/com/mirapy-org/mirapy.svg?style=flat-square
+.. image:: https://img.shields.io/travis/com/mirapy-org/mirapy.svg?style=flat-square&logo=travis%20ci
     :target: https://travis-ci.com/mirapy-org/mirapy
     :alt: Travis CI
 
@@ -21,11 +21,11 @@ MiraPy: Python Package for Deep Learning in Astronomy
     :target: https://coveralls.io/github/mirapy-org/mirapy
     :alt: Coveralls
 
-.. image:: https://img.shields.io/badge/chat%20on-Slack-E51670.svg?style=flat-square
+.. image:: https://img.shields.io/badge/chat%20on-Slack-4A154B.svg?style=flat-square&logo=slack
     :target: https://join.slack.com/t/mirapy/shared_invite/enQtNjEyNDQwNTI2NDY3LTE3ZmI3M2EyMjdkZWU4NTE2NjkxZjdhYWE4ZjUyODY0NzllNzRlMzZhNThhNWRiMjk4MjNhYWQ3NjA3YjJiNGY
     :alt: Slack
 
-.. image:: https://img.shields.io/pypi/v/mirapy.svg?style=flat-square
+.. image:: https://img.shields.io/pypi/v/mirapy.svg?style=flat-square&logo=pypi
     :target: https://pypi.org/project/mirapy/
     :alt: PyPI
 
