### 2019-07-27 #1557 @mbobra
@DanielLenz that is a good question.

We generally do not consider software that serve solely as wrappers, unless they add some additional functionality. If the submission does add something unique/different, that the underlying software does not do, then I think that could be in scope. The other criteria we sometimes apply is: would/should this software be cited if used in a paper? Or would the underlying software be the only thing that contributes significantly to the results?
