### 2019-07-29 #1557 @swapsha96
I used this software to beat state of the art results of different algorithms for image reconstruction of astronomical images ([tutorials](https://github.com/mirapy-org/tutorials/blob/master/Autoencoders/Astronomical%20Image%20Reconstruction%20using%20Autoencoder.ipynb)). This software has a CNN model that improves 'per-class accuracy' of pulsars from [78%](https://github.com/as595/HTRU1/blob/master/htru1_tutorial_channel.ipynb) to [91%](https://github.com/mirapy-org/tutorials/blob/master/Classifiers/HTRU1%20Batched%20Dataset%20Classification.ipynb). MiraPy can be used for transfer learning in classification of ATLAS variable star catalog which will be released in future. We have a curve fitting module (in initial stage) which uses `autograd` for faster gradient computation. These are few to mention as such. We will be bringing more features and improvements in future. Do you need more reasons for usage as such?
