### 2019-08-01 #1557 @DanielLenz
Thanks for the discussion, I'm curious to see what @wtgee thinks.

In the meantime, could you add references to all the Python libraries that you're using (not only in mirapy itself, but also in the tutorial)? Examples are sklearn, matplotlib, and numpy.

Furthermore, please add references (and DOIs, if they exist) to the data sets which you use in the tutorials.
