### 2019-08-01 #1557 @swapsha96
@DanielLenz a list of all Python libraries required for tutorials is [here](https://github.com/mirapy-org/tutorials/blob/master/requirements.txt).

Regarding, adding references, do you want me to add them (includind DOIs) in tutorials repository?
