### 2019-08-01 c2c33a2
Update paper.bib
