+++ paper.bib
@@ -4,3 +4,68 @@
   year={2015},
   howpublished={\url{https://keras.io}},
 }
+@article{Heinze_2018,
+	doi = {10.3847/1538-3881/aae47f},
+	url = {https://doi.org/10.3847%2F1538-3881%2Faae47f},
+	year = 2018,
+	month = {nov},
+	publisher = {American Astronomical Society},
+	volume = {156},
+	number = {5},
+	pages = {241},
+	author = {A. N. Heinze and J. L. Tonry and L. Denneau and H. Flewelling and B. Stalder and A. Rest and K. W. Smith and S. J. Smartt and H. Weiland},
+	title = {A First Catalog of Variable Stars Measured by the Asteroid Terrestrial-impact Last Alert System ({ATLAS})},
+	journal = {The Astronomical Journal},
+	abstract = {The Asteroid Terrestrial-impact Last Alert System (ATLAS) carries out its primary planetary defense mission by surveying about 13,000 deg2 at least four times per night. The resulting data set is useful for the discovery of variable stars to a magnitude limit fainter than r ∼ 18, with amplitudes down to 0.02 mag for bright objects. Here, we present a Data Release One catalog of variable stars based on analyzing the light curves of 142 million stars that were measured at least 100 times in the first two years of ATLAS operations. Using a Lomb–Scargle periodogram and other variability metrics, we identify 4.7 million candidate variables. Through the Space Telescope Science Institute, we publicly release light curves for all of them, together with a vector of 169 classification features for each star. We do this at the level of unconfirmed candidate variables in order to provide the community with a large set of homogeneously analyzed photometry and to avoid pre-judging which types of objects others may find most interesting. We use machine learning to classify the candidates into 15 different broad categories based on light-curve morphology. About 10% (427,000 stars) pass extensive tests designed to screen out spurious variability detections: we label these as “probable” variables. Of these, 214,000 receive specific classifications as eclipsing binaries, pulsating, Mira-type, or sinusoidal variables: these are the “classified” variables. New discoveries among the probable variables number 315,000, while 141,000 of the classified variables are new, including about 10,400 pulsating variables, 2060 Mira stars, and 74,700 eclipsing binaries.}
+}
+@article{1810.09489,
+Author = {A. Udalski and I. Soszyński and P. Pietrukowicz and M. K. Szymański and D. M. Skowron and J. Skowron and P. Mróz and R. Poleski and S. Kozłowski and K. Ulaczyk and K. Rybicki and P. Iwanek and M. Wrona},
+Title = {OGLE Collection of Galactic Cepheids},
+Year = {2018},
+Eprint = {arXiv:1810.09489},
+Howpublished = {2018, Acta Astron., 68, 315},
+Doi = {10.32023/0001-5237/68.4.1},
+}
+@article{Gopalan_2015,
+	doi = {10.1088/0004-637x/809/1/40},
+	url = {https://doi.org/10.1088%2F0004-637x%2F809%2F1%2F40},
+	year = 2015,
+	month = {aug},
+	publisher = {{IOP} Publishing},
+	volume = {809},
+	number = {1},
+	pages = {40},
+	author = {Giri Gopalan and Saeqa Dil Vrtilek and Luke Bornn},
+	title = {{CLASSIFYING} X-{RAY} {BINARIES}: A {PROBABILISTIC} {APPROACH}},
+	journal = {The Astrophysical Journal},
[…]
