### 2019-08-01 #1557 @DanielLenz
Thanks for adding the DOIs for the data sets!
I think it would be appropriate to also cite the different python packages in the paper, many of them have accompanying publications or other bibtex items (e.g. [for sklearn](https://scikit-learn.org/stable/about.html#citing-scikit-learn)).
