### 2019-08-09 d37cb6c
Update paper.md
