### 2019-08-09 #1557 @swapsha96
@mbobra @DanielLenz @whedon I've added citations of different Python packages in the paper.
I see one check is left in Daniel's list. So what's next?
