### 2019-08-10 #1557 @mbobra
👋@openjournals/joss-eics, @wtgee is no longer available to review this submission. Should I assign another reviewer or move forward with only one review of the submission?
