### 2019-08-22 #1557 @mbobra
👋 @swapsha96 I'm looking through the General Checks, which asks the following question:

> Has the submitting author (@swapsha96) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

It looks like @akhilsinghal1234 made most of the contributions to the software -- I only see [six commits](https://github.com/mirapy-org/mirapy/graphs/contributors) by you (all on the JOSS paper) to the master branch. Do you agree?
