### 2019-08-22 #1557 @swapsha96
No! I don't why total number of commits are [205](https://github.com/mirapy-org/mirapy/commits/master) but it says He made 45 commits and I made 6 commits. I started this project and Github says I wrote only 150 lines. WTH! 😆 Does anyone know why lots of my commits are not considered in contributor section? (Below is an screenshot)

(I can't believe I sound so arrogant here 😅)
