### 2019-08-22 #1557 @mbobra
I see. The contributor section counts commits by Github users, which are @swapsha96 and @akhilsinghal1234. It looks like there were other commits under the name Swapnil Sharma, which is not associated with a Github account. So those commits must have been via some other mechanism.
