### 2019-08-22 #1557 @kyleniemeyer
This may have happened due to a local Git user/email configuration that didn't match the GitHub account—it looks strange at first glance, but I don't think this is a major issue, since it's clear upon closer inspection that @swapsha96 is a major contributor. Good catch though, @mbobra!
