### 2019-08-23 #1557 @mbobra
@swapsha96 I spent the afternoon playing with MiraPy. I agree with @DanielLenz that it is essentially a shallow wrapper around Keras. Therefore, this does not meet our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) (it falls under the 'thin wrapper' category).

However, I do think MiraPy provides excellent tutorials on how to use machine learning to accomplish the tasks outlined on the tutorials page. I would strongly recommend this for publication in [JOSE](https://jose.theoj.org/).

cc @openjournals/joss-eics
